Linked comments to users

+++ src/Controller/CommentController.php
@@ -9,12 +9,19 @@ use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\Request;
 use Symfony\Component\HttpFoundation\Response;
 use Symfony\Component\Routing\Annotation\Route;
+use Symfony\Component\Security\Core\Authorization\AuthorizationCheckerInterface;
+use Symfony\Component\Security\Http\Attribute\IsGranted;
 
 class CommentController extends AbstractController
 {
     #[Route('/comment/{id}/edit', name: 'comment_edit')]
+    #[IsGranted("ROLE_USER")]
     public function edit(Comment $comment, Request $request, EntityManagerInterface $entityManager): Response
     {
+        if ($this->getUser() !== $comment->getPoster()) {
+            return $this->redirectToRoute('homepage');
+        }
+
         $form = $this->createForm(CommentType::class, $comment);
 
         $form->handleRequest($request);
@@ -32,8 +39,12 @@ class CommentController extends AbstractController
     }
 
     #[Route('/comment/{id}/delete', name: 'comment_delete', methods: ["POST"])]
-    public function delete(Comment $comment, EntityManagerInterface $entityManager): Response
+    public function delete(Comment $comment, EntityManagerInterface $entityManager, AuthorizationCheckerInterface $authorizationChecker): Response
     {
+        if ($this->getUser() !== $comment->getPoster() && !$authorizationChecker->isGranted("ROLE_ADMIN")) {
+            return $this->redirectToRoute('homepage');
+        }
+
         $slug = $comment->getPost()->getSlug();
         $entityManager->remove($comment);
         $entityManager->flush();
